backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Config
pydantic-settings==2.2.1

# Video / Image processing
ffmpeg-python==0.2.0
Pillow==10.3.0
imagehash==4.3.1

# Vector search (CPU only, no GPU required)
faiss-cpu==1.8.0

# Phase 2: Watermarking (CPU-only DWT-DCT-SVD)
invisible-watermark==0.2.0
opencv-python-headless==4.9.0.80

# Phase 3: URL monitoring / video download
yt-dlp==2024.5.27

# Phase 7: Deep learning embeddings (CPU)
# torch + torchvision are installed separately in Dockerfile via PyTorch CPU wheel index
# to avoid pulling the large CUDA build (~2 GB). Do NOT add them here.
onnxruntime==1.18.0

# System monitoring (for /api/system-info)
psutil==5.9.8

# Google AI (Gemini)
google-generativeai==0.4.0

# Utilities
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
